Package updater - separate local from github ops This is step towards having those as a separate lib / separately invokable
