+++ NuKeeper/Engine/Packages/PackageUpdater.cs
@@ -1,12 +1,9 @@
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
-using NuKeeper.Configuration;
 using NuKeeper.Git;
 using NuKeeper.Github;
 using NuKeeper.Inspection.Logging;
 using NuKeeper.Inspection.RepositoryInspection;
-using NuKeeper.NuGet.Process;
 using Octokit;
 
 namespace NuKeeper.Engine.Packages
@@ -15,16 +12,16 @@ namespace NuKeeper.Engine.Packages
     {
         private readonly IGithub _github;
         private readonly INuKeeperLogger _logger;
-        private readonly UserSettings _settings;
+        private readonly ILocalPackageUpdater _localUpdater;
 
         public PackageUpdater(
             IGithub github,
-            INuKeeperLogger logger,
-            UserSettings settings)
+            ILocalPackageUpdater localUpdater,
+            INuKeeperLogger logger)
         {
             _github = github;
+            _localUpdater = localUpdater;
             _logger = logger;
-            _settings = settings;
         }
 
         public async Task UpdatePackageInProjects(
@@ -43,7 +40,7 @@ namespace NuKeeper.Engine.Packages
                 _logger.Verbose($"Using branch name: '{branchName}'");
                 git.CheckoutNewBranch(branchName);
 
-                await UpdateAllCurrentUsages(updateSet);
+                await _localUpdater.Update(updateSet);
 
                 var commitMessage = CommitWording.MakeCommitMessage(updateSet);
                 git.Commit(commitMessage);
@@ -58,47 +55,6 @@ namespace NuKeeper.Engine.Packages
             catch (Exception ex)
             {
                 _logger.Error("Update failed", ex);
-            }
-        }
-
-        private async Task UpdateAllCurrentUsages(PackageUpdateSet updateSet)
-        {
-            foreach (var current in updateSet.CurrentPackages)
-            {
-                var updateCommands = GetUpdateCommands(current.Path.PackageReferenceType);
-                foreach (var updateCommand in updateCommands)
-                {
-                    await updateCommand.Invoke(updateSet.SelectedVersion, updateSet.Selected.Source, current);
-                }
-            }
-        }
-
-        private IReadOnlyCollection<IPackageCommand> GetUpdateCommands(PackageReferenceType packageReferenceType)
-        {
-            switch (packageReferenceType)
-            {
-                case PackageReferenceType.PackagesConfig:
-                    return new IPackageCommand[]
-                    {
-                        new NuGetFileRestoreCommand(_logger, _settings),
-                        new NuGetUpdatePackageCommand(_logger, _settings)
-                    };
-
-                case PackageReferenceType.ProjectFileOldStyle:
-                    return new IPackageCommand[]
-                    {
-                        new UpdateProjectImportsCommand(),
-                        new NuGetFileRestoreCommand(_logger, _settings),
-                        new DotNetUpdatePackageCommand(_logger, _settings)
-                    };
-
-                case PackageReferenceType.ProjectFile:
-                    return new[] {new DotNetUpdatePackageCommand(_logger, _settings)};
-
-                case PackageReferenceType.Nuspec:
-                    return new[] { new UpdateNuspecCommand(_logger) };
-
-                default: throw new ArgumentOutOfRangeException(nameof(packageReferenceType));
             }
         }
